Revert "simplifications"
This reverts commit 7cbe4e8e5bed056fd6218885b8d69d4998882ff5.

# Conflicts:
#	ManagedDoom/src/Doom/Game/DoomGame.cs
#	ManagedDoom/src/Doom/Game/SaveAndLoad.cs

+++ ManagedDoom/src/Video/ThreeDRenderer.cs
@@ -2415,7 +2415,7 @@ namespace ManagedDoom.Video
             {
                 // Choose a different rotation based on player view.
                 var ang = Geometry.PointToAngle(viewX, viewY, thingX, thingY);
-                var rot = (ang.Data - thing.Angle.Data + Angle.Ang45.Data / 2 * 9) >> 29;
+                var rot = (ang.Data - thing.Angle.Data + (uint)(Angle.Ang45.Data / 2) * 9) >> 29;
                 lump = spriteFrame.Patches[rot];
                 flip = spriteFrame.Flip[rot];
             }
